// src/app/core/models/order.model.ts
export class OrderModel {
    _id?: string;
    name?: string;
    description?: string;
    type?: string;
    createdAt?: Date;
    updatedAt?: Date;
    userName?: string;
    grabitValue?: number;
    userId?: string;
    desc?: string;
    email?: string;
    city?: string;
    phone?: string;
    age?: string;
    attachments?: Array<any>;
    isVerified?: boolean;
    isLock?: boolean;
    isDeleted?: boolean;
    isAdmin?: boolean;
    profileUrl?: string;
}